requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.4
coverage==6.4.1
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
djlint==1.1.1
flake8==4.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.11.4
Markdown==3.3.7
mccabe==0.6.1
pathspec==0.9.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyYAML==6.0
regex==2022.6.2
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
zipp==3.8.0
Pillow==9.1.1
